fix: foreign key question in question vote table

+++ internal/models/question.go
@@ -3,7 +3,7 @@ package models
 import "time"
 
 type Question struct {
-	Id        int64     `gorm:"primaryKey" json:"id"`
+	Id        uint      `gorm:"primaryKey" json:"id"`
 	UserName  string    `gorm:"type:varchar(255)" json:"user_name"`
 	Question  string    `gorm:"type:varchar(255)" json:"question"`
 	Answers   []Answer  `gorm:"foreignKey:QuestionId" json:"answers"`
